requirements.txt
PyGithub==2.1.1
rich==13.7.0
python-dotenv==1.0.0